+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/view/FavoriteFragment.kt
@@ -1,4 +1,4 @@
-package com.wonchihyeon.ytt_android.fragments
+package com.wonchihyeon.ytt_android.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,14 +6,13 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import androidx.navigation.findNavController
 import com.wonchihyeon.ytt_android.R
 import com.wonchihyeon.ytt_android.databinding.FragmentFavoriteBinding
 
 
 
 // 즐겨찾기 페이지
-class FavoriteFragment : Fragment() {
+class FavoriteFragment : Fragment(){
 
     private lateinit var binding: FragmentFavoriteBinding
