+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/view/order/OrderItem.kt
@@ -1,4 +1,4 @@
-package com.wonchihyeon.ytt_android
+package com.wonchihyeon.ytt_android.view.order
 
 data class OrderItem(val name: String, val basePrice: String, val totalPrice: String)
