+++ FE/app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".account.LoginActivity">
+        tools:context=".view.account.LoginActivity">
 
         <LinearLayout
             android:layout_marginTop="60dp"
