+++ FE/app/src/main/res/layout/activity_main.xml
@@ -28,5 +28,7 @@
             app:itemTextColor="@drawable/bottom_navigation_color"
             app:itemIconTint="@drawable/bottom_navigation_color"/>
 
+
+
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
